+++ transaction-service/src/main/java/cz/muni/fi/obs/api/TransactionCreateDto.java
@@ -1,5 +1,7 @@
 package cz.muni.fi.obs.api;
 
+import java.math.BigDecimal;
+
 import jakarta.validation.constraints.Min;
 import jakarta.validation.constraints.NotBlank;
 import lombok.Builder;
@@ -10,8 +12,8 @@ public record TransactionCreateDto(
 		String withdrawsFrom,
 		@NotBlank
 		String depositsTo,
-		@Min(0)long withdrawAmount,
-		@Min(0)long depositAmount,
+		@Min(0) BigDecimal withdrawAmount,
+		@Min(0) BigDecimal depositAmount,
 		String note,
 		String variableSymbol) {
 }
